refactor: extract Program entity from Recorded
